I grabbed a Radix sort from the web that is not only smaller than mine but more efficient than any of the sorting algorithms I have covered. I don't totally understand it yet.
